api/requirements.txt
Django==2.2.3
django-cors-headers==3.0.2
graphene==2.1.7
graphene-django==2.4.0
graphql-core==2.2.1
graphql-relay==2.0.0
promise==2.2.1
pytz==2019.1
singledispatch==3.4.0.3
six==1.12.0
sqlparse==0.3.0
django-jsonfield==1.4.0